move categories array into assets folder
