+++ client/src/assets/categories.js
